src/functions/analysis/main.py
import os
import json
import snowflake.connector
from flask import escape

def analyze_data(request):
    """HTTP Cloud Function.
    Args:
        request (flask.Request): The request object.
    Returns:
        The response text, or any set of values that can be turned into a
        Response object using `make_response`
    """
    request_json = request.get_json(silent=True)
    
    # In a real scenario, we would get credentials from Secret Manager
    # For this prototype, we'll assume they are passed or env vars
    # snowflake_account = os.environ.get('SNOWFLAKE_ACCOUNT')
    # snowflake_user = os.environ.get('SNOWFLAKE_USER')
    # snowflake_password = os.environ.get('SNOWFLAKE_PASSWORD')

    print("Connecting to Snowflake...")
    # ctx = snowflake.connector.connect(
    #     user=snowflake_user,
    #     password=snowflake_password,
    #     account=snowflake_account
    #     ...
    # )
    # cs = ctx.cursor()
    # cs.execute("SELECT cortex_predict(...) FROM data")
    # results = cs.fetchall()
    
    # Mock Result
    results = [
        {"id": "1", "prediction": "High Growth", "confidence": 0.95},
        {"id": "2", "prediction": "Stable", "confidence": 0.88}
    ]

    print("Analysis complete.")
    return json.dumps({"status": "success", "data": results}), 200, {'Content-Type': 'application/json'}
