src/functions/analysis/requirements.txt
snowflake-connector-python
flask
